Add DM command to 'recollect' assets at runtime
Testing new faces by comparing old faces with new ones is annoying
because it requires rebooting the server. Make this easier by adding a
command to cause the server to run the collection process again.

Don't document this just yet, since this command not might be suitable
for production servers just yet.