Clear faces_sent if client disconnects early
faces_sent is used to keep track of faces that have been sent to a
client. The same dynamically-allocated faces_sent is used in the
pre-login init_socket and then copied to the player sock_struct after
logging in.

faces_sent is not correctly cleared if a pre-login client has faces sent
but disconnects. Fix this by always using calloc() in init_connection()
and freeing in the do_server() pre-login connection teardown.

Tracker:    https://sourceforge.net/p/crossfire/bugs/783/