Refactor shop pricing
Modify shop pricing to be in terms of an "efficiency" between 0-1. The
efficiency modifies the base price on both buy and sell. It is
influenced by shop greed, approval rating, and charisma.

Change specialization to only apply to items sold to shops. Remove a few
manual hacks. Add CF_DEBUG_SHOP environment variable.

Due to changes in charisma and greed handling, shopkeepers are now
slightly less greedy (items sell for more, and it costs less to buy
items back).