Add object load/save test
This test loads a map given to it on the command line, then saves the
objects without map headers to a temporary file. This works around
object field order inconsistency from Gridarta or older servers.

Then, it loads objects from the temporary file, saves them to a second
temporary file, and runs diff to compare them. Ideally, the two files
should be identical.

Note that this test by itself is not sufficient to check that map
load/save is working correctly. For example, this test would still pass
if save_objects() always produces a blank file.

This test currently should only be run from the command line.