Give non-moving objects a move type in move_ob()
move_ob() is sometimes called on inanimate objects without a move type,
for example, as an effect of check_spell_knockback(). As a result, these
objects skip move_on event checking, for example, bypassing button
presses and releases.

Fix this by temporarily giving objects without a move type MOVE_WALK,
which is necessary to trigger buttons and such.

It may be better in the future to warn about missing move_type when
calling move_ob(), so that the calling code is responsible for setting
and resetting move_ob.