Use priority queue for active list processing
Based on profiling data, the server spends a lot of time decrementing
the speed_left variable for every active object each tick.

This change avoids needing to every speed_left each tick by computing
when an object is expected to be active again, and pushing it into a
priority queue. speed_left is checked and then updated only as each
object is expected to become active.