Make player socket a pointer
Change the socket in the player struct to a pointer so that it can be
passed around independently of the player struct.

In the beginning, sockets are still created from init_sockets. When an
init_socket becomes a player, space for a new socket is allocated using
malloc() and copied from init_sockets using memcpy(), then stored as a
pointer into the player struct. Storage for the player socket is freed
when the client disconnects.