Set TCP_NODELAY when handling player commands
For the most part, the server keeps TCP_NODELAY off because socklists
send a lot of little packets that should be coalesced according to
Nagle's algorithm to reduce TCP overhead.

However, when sending responses to a player command, Nagle's algorithm
allows only the first socklist response to be transmitted immediately.
Subsequent socklists, including the comc command acknowledgement, are
delayed for coalescing. This adds latency to the subsequent packets.

The best solution is, of course, to call send() only once after putting
together all of the server response socklists. This, however, would
require significant rework on the socklists.

For now, set TCP_NODELAY when starting to handle a player command and
turn it back off when the comc is sent.

While here, remove the TCP_NODELAY twiddle from the tick command. The
tick command isn't actually latency sensitive and so can be coalesced
when possible without the client noticing anything.