Truncate when exceeding MAXSOCKBUF
Sometimes SockList_AddPrintf() is called with an enormous argument that
overruns MAXSOCKBUF, which currently crashes the server. Instead of
crashing, just truncate the message. This will likely cause the next
message to be corrupt and cause the client to disconnect, but this is
better than crashing the game for everyone.