Remove debug print
It is quite possible for the game to not find a enemy even if the square
is alive, e.g. a pet's owner is standing on that square.

Remove this debug log because with lots of pets this can severely slow
down a server with debug logging to a disk.