Cache monster_stand_in_light()
The algorithm is very slow and leads to performance problems on
invidious when there are a lot of monsters and a lot of pets.

Pave over the problem by memoizing the computation, so that it is not
duplicated for each monster that is looking for a new player/friendly to
target.