Remove block_until_new_connection()
This has been broken for some time without anyone noticing because on
modern systems there are two listening sockets (one for IPv4 and one for
IPv6) and so this has never been called.

In any case it is still broken because when a client connects, the first
select() in block_until_new_connection() returns a positive number, but
then the second select() in do_server() returns zero since we didn't
accept().

I'm looking at fixing this but for now just remove the old broken code.