Disable create_home_portals
create_home_portals allows players to create town portals inside
player-unique maps. The problem with these is that the portals close
whenever the unique map is swapped, leading the portals to close
prematurely.

This is being disabled because server operators are inadvertently
enabling this, then complaining about town portals being broken.

This can be brought back when someone investigates and implements a fix
for keeping town portals open when one end is in a swapped player-unique
map.