Remove custom map swap times
Maps used to be able to specify how often they wanted to be swapped.
This has been broken for some time so remove it.

This allows us to reset the swap timeout on a map whenever it is
referenced by ready_map_name() instead of on load via get_linked_map().
This will hopefully let us make better decisions on which maps to
prioritize swapping.

This should also fix an issue with town portals, where creating a town
portal on a map that is swapped creates a portal that appears to be
"closed" because the other end was unswapped for one tick (via
ready_map_name) but is instantly swapped out the next tick before the
portal opening animation is finished.