Require shopmax to be set to buy items from players
Previously, shops with shopmax set to zero (the default) or unset had no
limit for the price that shops would pay to buy items from players.

To allow some shops to be "buy only", this changes the zero setting to
indicate that the shop will not buy items from players.