Increase delay buffer counter width
Several recent crashes on Invidious suggest that delayed_buffers_used or
delayed_buffers_allocated are overflowing, causing sl to be set to an
empty or invalid SockList and crashing the server with a segfault.

So for now, increase the width of the integer to try to prevent crashes
while someone investigates the root cause.