Fix character creation after deleting with same name
If you delete a character, then immediately create one with the same
name without playing a different character or disconnecting first, you
will receive an incorrect error message about the character existing.

This is because fields in op are not cleared until the client plays a
different character or disconnects. Therefore, after deleting a
character, immediately creating a new character with the same name as
the deleted one fails because verify_player() thinks this player is
still playing.

Hack around this by clearing the name right after deletion.