Enable shared library building
Build convenience libraries as libtool libraries, so that building as
shared libraries is enabled on platforms that support them.

While here, eliminate libsocket and libtypes, and include that in
libserver. These two libraries are not really useful on their own
without also linking to the server. This also improves build
parallelization.