Allow players to reset quest state
A big complaint about the new quest system is that players get their
quest state permanently stuck when they lose quest items or they get
quest items in the wrong order.

By popular demand, allow players to reset their own quest progress. This
effectively makes all quests restartable. But it is easier to allow this
than it is to go and fix all the quests and make sure every quest has
robust contingency plans built in.

To support this, allow a state == 0 argument to quest_set_state(). This
models a state where the clients know that a quest exists, but the
character has made no progress on this quest. This is required because
there is basically no delquest client command. In the future, this can
be used to support a server making a list of all quests available, even
if they are not started.